+++ OpenGL/GameEngine/GameEngine/boundingSphere.h
@@ -11,7 +11,7 @@ class BoundingSphere : public Collider
 {
 public:
 
-	BoundingSphere(const Vector3f& center, float radius) :
+	BoundingSphere(const Vector3f& center, double radius) :
 		Collider(Collider::TYPE_SPHERE),
 		m_center(center),
 		m_radius(radius) {}
@@ -21,11 +21,11 @@ public:
 	virtual void Transform(const Vector3f& translation);
 	virtual Vector3f GetCenter() const { return m_center; }
 
-	inline float GetRadius()           const { return m_radius; }
+	inline double GetRadius()           const { return m_radius; }
 
 private:
 	Vector3f m_center;
-	float    m_radius;
+	double    m_radius;
 };
 
 #endif // BOUNDINGSPHERE_H
