+++ OpenGL/GameEngine/GameEngine/physicsEngine.cpp
@@ -6,7 +6,7 @@ void PhysicsEngine::AddObject(const PhysicsObject& object)
 	m_objects.push_back(object);
 }
 
-void PhysicsEngine::Simulate(float delta)
+void PhysicsEngine::Simulate(double delta)
 {
 	for(unsigned int i = 0; i < m_objects.size(); i++)
 	{
